Feature/delete account (#27)
* Fixing WKHTMLTOPDF binary on Windows, starting to add logic for delete account only when all recurring donations have been cancelled

* Do not delete a user who has recurring donations un-cancelled

* Fixed delete user to only work if there are no recurring payments, added more test cases, etc.

+++ app/controllers/users_controller.rb
@@ -19,8 +19,6 @@ class UsersController < ApplicationController
         @user = User.new(user_params)
         
         if @user.save(context: current_admin ? :admin : :user)
-            redirect_to recurring_donation_transactions_path(from: 'create') and return unless @user.membership == "" || @user.membership.nil?
-            
             if current_admin
                 redirect_to user_path(current_admin.id) and return
             else
@@ -112,6 +110,13 @@ class UsersController < ApplicationController
         end
 
         @user = User.find(params[:id])
+
+        unless @user.made_donations.where(recurring: true).empty?
+            flash[:notice] = "Please delete your recurring donations before deleting your account."
+            flash.keep
+            render 'edit' and return
+        end
+
         if @user.destroy
             redirect_to (is_currently_admin?) ? users_path : user_path(current_user.id) and return
         else
@@ -165,9 +170,9 @@ class UsersController < ApplicationController
 
     def user_params
         if current_admin
-            params.require("user").permit(:email, :password, :password_confirmation, :membership, :admin)
+            params.require("user").permit(:email, :password, :password_confirmation, :admin)
         else
-            params.require("user").permit(:first_name, :last_name, :email, :password, :password_confirmation, :street_address_line_1, :city, :state, :zip_code, :street_address_line_2, :membership)
+            params.require("user").permit(:first_name, :last_name, :email, :password, :password_confirmation, :street_address_line_1, :city, :state, :zip_code, :street_address_line_2)
         end
     end
